Move coordinator to separate file for freedompro
