+++ homeassistant/components/freedompro/__init__.py
@@ -1,19 +1,15 @@
 """Support for freedompro."""
 from __future__ import annotations
 
-from datetime import timedelta
 import logging
-from typing import Any, Final
-
-from pyfreedompro import get_list, get_states
+from typing import Final
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_API_KEY, Platform
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import aiohttp_client
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import DOMAIN
+from .coordinator import FreedomproDataUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -58,39 +54,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 async def update_listener(hass: HomeAssistant, config_entry: ConfigEntry) -> None:
     """Update listener."""
     await hass.config_entries.async_reload(config_entry.entry_id)
-
-
-class FreedomproDataUpdateCoordinator(DataUpdateCoordinator[list[dict[str, Any]]]):
-    """Class to manage fetching Freedompro data API."""
-
-    def __init__(self, hass, api_key):
-        """Initialize."""
-        self._hass = hass
-        self._api_key = api_key
-        self._devices: list[dict[str, Any]] | None = None
-
-        update_interval = timedelta(minutes=1)
-        super().__init__(hass, _LOGGER, name=DOMAIN, update_interval=update_interval)
-
-    async def _async_update_data(self):
-        if self._devices is None:
-            result = await get_list(
-                aiohttp_client.async_get_clientsession(self._hass), self._api_key
-            )
-            if result["state"]:
-                self._devices = result["devices"]
-            else:
-                raise UpdateFailed()
-
-        result = await get_states(
-            aiohttp_client.async_get_clientsession(self._hass), self._api_key
-        )
-
-        for device in self._devices:
-            dev = next(
-                (dev for dev in result if dev["uid"] == device["uid"]),
-                None,
-            )
-            if dev is not None and "state" in dev:
-                device["state"] = dev["state"]
-        return self._devices
